guest_genderの採番ルール変更

+++ database/migrations/2021_06_28_103507_create_guest_genders_table.php
@@ -14,7 +14,7 @@ class CreateGuestGendersTable extends Migration
     public function up()
     {
         Schema::create('guest_genders', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->Integer('id');
             $table->string('guest_gender');
             $table->timestamps();
         });
